refactor: fit the fallthrough suites under the source budget
The exact-oracle pins pushed the CLI table to 371 and the source
threading pushed the generator to 352; the case table moves to a sibling
module and the expression helpers to a #[path] submodule, both files back
under 350 with behavior untouched.
